roachtest: move perturbation tests to own package
This commit moves the perturbation tests to their own package. It
doesn't change anything about the tests.

Release note: None
