+++ pkg/cmd/roachtest/tests/BUILD.bazel
@@ -19,7 +19,6 @@ go_library(
         "admission_control_index_backfill.go",
         "admission_control_index_overload.go",
         "admission_control_intent_resolution.go",
-        "admission_control_latency.go",
         "admission_control_multi_store_overload.go",
         "admission_control_multitenant_fairness.go",
         "admission_control_row_level_ttl.go",
@@ -230,6 +229,7 @@ go_library(
         "//pkg/cmd/roachtest/roachtestutil/task",
         "//pkg/cmd/roachtest/spec",
         "//pkg/cmd/roachtest/test",
+        "//pkg/cmd/roachtest/tests/perturbation",
         "//pkg/gossip",
         "//pkg/internal/sqlsmith",
         "//pkg/internal/workloadreplay",
@@ -285,7 +285,6 @@ go_library(
         "//pkg/util/uuid",
         "//pkg/util/version",
         "//pkg/workload",
-        "//pkg/workload/cli",
         "//pkg/workload/debug",
         "//pkg/workload/histogram",
         "//pkg/workload/histogram/exporter",
